Rebrands package under com.teragrep
