--- src/test/java/com/tersesystems/jmxbuilder/ParameterInfoTests.java
@@ -1,4 +1,0 @@
-package com.tersesystems.jmxbuilder;
-
-public class ParameterInfoTests {
-}
