--- src/test/java/com/tersesystems/jmxbuilder/TabularDataTests.java
@@ -1,5 +1,0 @@
-package com.tersesystems.jmxbuilder;
-
-public class TabularDataTests {
-
-}
