--- src/test/java/com/tersesystems/jmxbuilder/model/ExampleService.java
@@ -1,17 +1,0 @@
-package com.tersesystems.jmxbuilder.model;
-
-public class ExampleService {
-    private boolean debug;
-
-    public String dump() {
-        return ("Dumping contents");
-    }
-
-    public boolean isDebugEnabled() {
-        return debug;
-    }
-
-    public void setDebugEnabled(boolean debug) {
-        this.debug = debug;
-    }
-}
